Requested refactor
