+++ api/src/models/user_progress/__init__.py
@@ -1,4 +1,4 @@
 from .table import UserProgress, AssignmentStatus
-from .schemas import CertificateDTO
+from .schemas import CertificateDTO, ProgressUpdate, SectionComplete
 
-__all__ = ["UserProgress", "AssignmentStatus", "CertificateDTO"]
+__all__ = ["UserProgress", "AssignmentStatus", "CertificateDTO", "ProgressUpdate", "SectionComplete"]
