+++ api/src/models/user_progress/schemas.py
@@ -2,6 +2,16 @@ from typing import Optional
 from sqlmodel import SQLModel
 
 
+class ProgressUpdate(SQLModel):
+  section_id: str
+  task_id: str
+
+
+class SectionComplete(SQLModel):
+  section_id: str
+  total_sections: int
+
+
 class CertificateDTO(SQLModel):
     user_id: str
     course_id: str
